[FeeEscalationDisabling] Fee values promoted to uint_64. NOT TESTED YET

+++ src/casinocoin/ledger/ReadView.h
@@ -52,9 +52,9 @@ namespace casinocoin {
 struct Fees
 {
     std::uint64_t base = 0;         // Reference tx cost (drops)
-    std::uint32_t units = 0;        // Reference fee units
-    std::uint32_t reserve = 0;      // Reserve base (drops)
-    std::uint32_t increment = 0;    // Reserve increment (drops)
+    std::uint64_t units = 0;        // Reference fee units
+    std::uint64_t reserve = 0;      // Reserve base (drops)
+    std::uint64_t increment = 0;    // Reserve increment (drops)
 
     Fees() = default;
     Fees (Fees const&) = default;
